# iris-cli/docs/openapi.complete.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "iris Complete API",
    "version": "2.0.0",
    "description": "Complete API specification for iris Dashboard (port 4319), iris-lead (port 5010), Vibe Studio (port 4320), and iris-cli (variable port). Includes all agent orchestration, messaging integrations, memory management, and system control endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4319",
      "description": "Dashboard API (primary web interface)"
    },
    {
      "url": "http://127.0.0.1:5010",
      "description": "iris-lead API (orchestration and chat)"
    },
    {
      "url": "http://127.0.0.1:4320",
      "description": "Vibe Studio API (editor and sessions)"
    },
    {
      "url": "http://127.0.0.1:4321",
      "description": "iris-cli API (agent router and tasks)"
    }
  ],
...
  "paths": {
...
    "/api/rag/index": {
      "post": {
        "tags": [
          "RAG"
        ],
        "summary": "POST /api/rag/index",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
...
}